split command classes into separate files
