+++ docs/source/api/datatypes.rst
@@ -1,0 +1,5 @@
+Datatypes docs
+====================
+
+.. automodule:: flightplandb.datatypes
+   :members:
