+++ docs/source/api/flightplandb.rst
@@ -1,0 +1,15 @@
+FlightplanDB docs
+====================
+
+.. automodule:: flightplandb.flightplandb
+   :members:
+   :undoc-members:
+
+Submodules
+^^^^^^^^^^^^^^^^^^^^^
+.. toctree::
+   :titlesonly:
+   :glob:
+   :maxdepth: 4
+
+   submodules/*
