+++ docs/source/api/submodules/nav.rst
@@ -1,0 +1,6 @@
+Nav docs
+====================
+
+.. automodule:: flightplandb.submodules.nav
+   :members:
+   :undoc-members:
