+++ docs/source/api/submodules/plan.rst
@@ -1,0 +1,6 @@
+Plan docs
+====================
+
+.. automodule:: flightplandb.submodules.plan
+   :members:
+   :undoc-members:
