+++ docs/source/api/submodules/tags.rst
@@ -1,0 +1,6 @@
+Tags docs
+====================
+
+.. automodule:: flightplandb.submodules.tags
+   :members:
+   :undoc-members:
