+++ docs/source/api/submodules/user.rst
@@ -1,0 +1,6 @@
+User docs
+====================
+
+.. automodule:: flightplandb.submodules.user
+   :members:
+   :undoc-members:
