+++ docs/source/api/submodules/weather.rst
@@ -1,0 +1,6 @@
+Weather docs
+=====================
+
+.. automodule:: flightplandb.submodules.weather
+   :members:
+   :undoc-members:
